+++ packages/app/hivecommand-frontend/src/views/Editor/pages/tags/import-modal/import.tsx
@@ -23,6 +23,7 @@ export const ImportView = (
                 type: string
             }[]
         }[];
+       
         onChange: (importMap: any) => void 
     }) => {