+++ packages/app/hivecommand-native/src/views/setup/stages/discoveryServer.tsx
@@ -7,7 +7,7 @@ export const DiscoveryServerStage = () => {
     const { state, setState } = useContext(SetupContext);
 
     return (
-        <Box sx={{flex: 1, display: 'flex', paddingLeft: '6px', paddingRight: '6px', alignItems: 'center'}}>
+        <Box sx={{flex: 1, display: 'flex', paddingLeft: '6px', paddingRight: '6px', alignItems: 'center', justifyContent: 'center', flexDirection: 'column'}}>
             <TextField 
                 value={state.discoveryServer || ''}
                 onChange={(e) => setState('discoveryServer', e.target.value)}
@@ -15,6 +15,15 @@ export const DiscoveryServerStage = () => {
                 fullWidth 
                 label="Discovery Server" 
                 />
+
+                
+            <TextField 
+                sx={{marginTop: '12px'}}
+                value={state.provisionCode || ''}
+                onChange={(e) => setState('provisionCode', e.target.value)}
+                label="Provisioning code" 
+                fullWidth 
+                size="small" />
 
                 
         </Box>