+++ packages/app/hivecommand-native/src/views/setup/stages/index.tsx
@@ -1,4 +1,3 @@
 export * from './discoveryServer'
-export * from './provisionCode'
 export * from './opcuaServer'
 export * from './layoutDownload'