+++ packages/core/command-data/prisma/migrations/20230601001121_add_optional_tag_scope/migration.sql
@@ -1,0 +1,5 @@
+-- AlterTable
+ALTER TABLE "ProgramTag" ADD COLUMN     "scopeId" TEXT;
+
+-- AddForeignKey
+ALTER TABLE "ProgramTag" ADD CONSTRAINT "ProgramTag_scopeId_fkey" FOREIGN KEY ("scopeId") REFERENCES "ProgramDataScope"("id") ON DELETE SET NULL ON UPDATE CASCADE;