+++ packages/app/hivecommand-native/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.15-alpha.221](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.220...v1.4.15-alpha.221) (2023-06-01)
+
+**Note:** Version bump only for package @hive-command/native
+
+
+
+
+
+## [1.4.15-alpha.220](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.219...v1.4.15-alpha.220) (2023-06-01)
+
+**Note:** Version bump only for package @hive-command/native
+
+
+
+
+
+## [1.4.15-alpha.219](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.218...v1.4.15-alpha.219) (2023-05-31)
+
+**Note:** Version bump only for package @hive-command/native
+
+
+
+
+
+## [1.4.15-alpha.218](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.217...v1.4.15-alpha.218) (2023-05-31)
+
+**Note:** Version bump only for package @hive-command/native
+
+
+
+
+
 ## [1.4.15-alpha.215](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.214...v1.4.15-alpha.215) (2023-05-15)
 
 **Note:** Version bump only for package @hive-command/native